add conditioned lifecycle checker
